Recommend consolidation refactor
